Refactor install script to do upgrades for v1.1
Closes #8

This includes a slight change for v1.1, as the post-receive hook is now
a symlink post-receive.sh. Previously, all post-receive hooks were
boilerplate and just called this script anyway.

If a user has modified the post-receive hook in any way, it will be left
untouched for the sake of compatability. Presumably most users won't
have done this though, so they get the benefit of using the symlinked
script.